Format source with Black
Gramps CI uses the latest available version of Black when checking formatting in GitHub Actions. At the moment Black is at version 24.4.2. To match the formatting of source with the CI system, use the same version.

There were two issues Black was calling out in this source. One at line 61 where it wanted to add an additional blank line between imports and the rest of the code, and the second on line 387 where there were spaces on an empty line which were removed.