Fixes bug #0013415. Correctly assign sortval = 0 when a date is EMPTY
The is_valid() method for a Date simply checks the sortval for a zero value. However, if an empty date is passed to _calc_sort_value() it first  adjusts to 1 ALL y,m,d values that are zero.  Thus, an empty date would return a sortval for the year 1-1-1 as a valid date.