Fixes bug #0013423.  Align unit test with fixes to Bug #0013415
This change to the unit-test code date_test.py is required so that the previous commit for bug #0013415 can pass the test.
The problematic part of the test calculates an age for somebody with a birth date of "" (an empty string) and a death date given by a year, such as 1760.
The previous test code required it to report "greater than 110 years", which should not be a suitable response. The previous code also reported an age of negative 1999 years for a birth year of 2000 and a death date of the empty string.
The unit test code now tests against a reply of "unknown".